Update Scan Cells
- Rename remaining s2cell files to scan cells
- Accounts for https://github.com/RealDeviceMap/RealDeviceMap/pull/388
